+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Remove hard dependency on `pkg` subfolder in app project folder
 * Report error when `install.sh` is run not from `build` folder
+* Add missing columns on write (upgrade schema) in DbFile
 
 ## 0.0.21 (2024-03-26)